backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
httpx==0.27.0
PyYAML==6.0.2
python-dotenv==1.0.1

# LiveKit
livekit==0.17.0
livekit-api==0.7.0

# LangGraph + LangChain
langgraph==0.2.0
langchain==0.3.0
langchain-core==0.3.0
langchain-openai==0.2.0

# Audio processing
numpy<2.0

# Redis (optional, for production session store)
redis[hiredis]==5.1.0
